+++ packages/apps/docs/src/components/Layout/Home/Home.tsx
@@ -1,34 +1,14 @@
-import { BaseTemplate, Footer, Header } from '../components';
+import { BasePageGrid } from '../components';
+import { Template } from '../components/Template';
 
 import { IPageProps } from '@/types/Layout';
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 
 export const Home: FC<IPageProps> = ({ children, leftMenuTree }) => {
-  const [isMenuOpen, setIsMenuOpen] = useState<boolean>(false);
-  const [isAsideOpen, setIsAsideOpen] = useState<boolean>(false);
-
-  const toggleMenu = (): void => {
-    setIsMenuOpen((v) => !v);
-    setIsAsideOpen(false);
-  };
-
-  const toggleAside = (): void => {
-    setIsAsideOpen((v) => !v);
-    setIsMenuOpen(false);
-  };
-
   return (
-    <BaseTemplate>
-      <Header
-        toggleMenu={toggleMenu}
-        toggleAside={toggleAside}
-        isMenuOpen={isMenuOpen}
-        isAsideOpen={isAsideOpen}
-        menuItems={leftMenuTree}
-      />
-      {children}
-      <Footer />
-    </BaseTemplate>
+    <BasePageGrid>
+      <Template menuItems={leftMenuTree}>{children}</Template>
+    </BasePageGrid>
   );
 };
