+++ packages/apps/docs/src/components/Layout/components/Template/index.ts
@@ -1,0 +1,1 @@
+export { Template } from './Template';
